requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain-community
langchain-text-splitters
faiss-cpu
sentence-transformers
groq
jinja2
python-multipart